+++ simulation/helpers.go
