+++ simulation/state.go
